parse_utils.c 생성, 함수 몇개 parse.c에서 분리시킴
