+++ parser/parser.c
@@ -1,73 +1,6 @@
 #include "minishell.h"
 
-int 	g_parse_error;
-
-int		count_token(char *input)
-{
-	int		count_token;
-	char	*p;
-
-	count_token = 1;
-	if (!(p = ft_calloc(ft_strlen(input) + 1, sizeof(char))))
-		return (0);
-	ft_strcpy(p, input);
-	if (ft_strtok(p, ' ') != NULL)
-	{
-		while (ft_strtok(NULL, ' ') != NULL)
-			count_token++;
-	}
-	free(p);
-	return (count_token);
-}
-
-void	*initialize(char *input, t_data *data, t_list **head)
-{
-	*head = ft_lstnew(NULL);
-	data->lstlast = *head;
-	if (!(data->cmd = ft_calloc(1, sizeof(t_cmd))))
-		return ((void *)parse_error(data, NULL, MALLOC_ERROR));
-	if (!(data->cmd->cmdline = ft_calloc(count_token(input) + 1, sizeof(char *)))) // 보류: char**이 calloc으로 가능?
-		return ((void *)parse_error(data, NULL, MALLOC_ERROR));
-	data->cmd->flag = 0;
-	data->cmd->quote = 0;
-	data->cmd->has_redir = 0; // 리다이렉션 플래그 초기화 추가
-	if (!(data->buff = ft_calloc(ft_strlen(input) + 1, sizeof(char))))
-		return ((void *)parse_error(data, NULL, MALLOC_ERROR));
-	data->i = -1;
-	data->j = 0;
-	data->k = 0;
-	return (NULL);
-}
-
-void	put_buff_into_cmdline(t_data *data)
-{
-	if (*(data->buff) == 0)
-		return ;
-	data->cmd->cmdline[(data->k)] = ft_strdup(data->buff);
-	data->cmd->cmdline[(data->k) + 1] = NULL;
-	(data->k)++;
-	ft_bzero(data->buff, ft_strlen(data->buff) + 1);
-	data->j= 0;
-}
-
-int		add_node(t_data *data, t_list *head, char *input, int symbol)
-{
-	data->cmd->flag = symbol;
-	if (*(data->buff))
-		put_buff_into_cmdline(data);
-	if ((data->cmd->cmdline)[0] == 0 && data->cmd->flag <= 1)
-		return (ERROR);
-	else
-	{
-		// head에 t_list *p(content:data->cmd, next:NULL)를 복사
-		ft_lstadd_back(&head, ft_lstnew(data->cmd));
-		data->cmd = ft_calloc(1, sizeof(t_cmd));
-		data->cmd->cmdline = ft_calloc(count_token(input) + 1, sizeof(char *));
-		data->lstlast = ft_lstlast(head);
-	}
-	data->k = 0;
-	return (SUCCESS);
-}
+int g_parse_error;
 
 void	exec_proc(t_list *head) // 인자는 연결리스트의 헤드포인터
 {
@@ -91,6 +24,36 @@ void	exec_proc(t_list *head) // 인자는 연결리스트의 헤드포인터
 		cur_proc = cur_proc->next;
 	}
 	ft_lstclear(&head, free_cmdline);
+}
+
+int		add_node(t_data *data, t_list *head, char *input, int symbol)
+{
+	data->cmd->flag = symbol;
+	if (*(data->buff))
+		put_buff_into_cmdline(data);
+	if ((data->cmd->cmdline)[0] == 0 && data->cmd->flag <= 1)
+		return (ERROR);
+	else
+	{
+		// head에 t_list *p(content:data->cmd, next:NULL)를 복사
+		ft_lstadd_back(&head, ft_lstnew(data->cmd));
+		data->cmd = ft_calloc(1, sizeof(t_cmd));
+		data->cmd->cmdline = ft_calloc(count_token(input) + 1, sizeof(char *));
+		data->lstlast = ft_lstlast(head);
+	}
+	data->k = 0;
+	return (SUCCESS);
+}
+
+void	put_buff_into_cmdline(t_data *data)
+{
+	if (*(data->buff) == 0)
+		return ;
+	data->cmd->cmdline[(data->k)] = ft_strdup(data->buff);
+	data->cmd->cmdline[(data->k) + 1] = NULL;
+	(data->k)++;
+	ft_bzero(data->buff, ft_strlen(data->buff) + 1);
+	data->j= 0;
 }
 
 void	parse_all_char(char *input, t_data *data, t_list *head)
@@ -142,27 +105,4 @@ void	*parse(char *input_temp)
 	if (g_parse_error == 1)
 		exec_proc(head);
 	return (NULL);
-}
-
-int		check_white_space(char *input)
-{
-	int	space;
-	int	i;
-
-	space = 0;
-	i = 0;
-	if (input[0] == 0)
-		return (1);
-	while (input[i])
-	{
-		if (input[i] == '\r' || input[i] == '\v'
-			|| input[i] == '\t' || input[i] == '\f')
-			return (1);
-		else if (input[i] == ' ')
-			space++;
-		i++;
-	}
-	if (space == (int)ft_strlen(input))
-		return (1);
-	return (0);
 }
